Use all subscription values between each epoch for calculating sub state
Using filtered-traces means that you will lose sub-state updates. The
filtered-traces are partial parts of an epoch, but it is only passed
to sub-state matching once there is a full epoch matched. At that point
filtered-epochs will not necessarily have a matching complement of
traces. This commit uses all of the stored traces on each run,
meaning that no traces will be dropped.