Properly handle popup windows being blocked on initial load
If our JS tries to open a popup window without user interaction, Chrome
will block this, as it is indistinguishable from a popup ad. To
reproduce this issue:

1. Ensure the popup window is closed
2. In Local Storage, set "day8.re-frame-10x.external-window?" => "true"
3. Reload the browser
4. re-frame-10x will try and re-open the popup to restore the state

If the popup window is already open, Chrome doesn't prevent you from
're-attaching' to it.

Fixing this exposed a latent bug where we try to reopen an external
window before we have dimensions loaded from localstorage, so this
commit fixes that issue too.

Fixes #228