Fix show/hide shortcut when using alternate keyboard layouts
We were previously using ev.code, which specified the location of the
key on a QWERTY layout. When using a different layout, 10x was listening
for a key press on the key which would be H under QWERTY.

This commit changes the listener to use ev.key which is keyboard layout
independent.

https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/key

Fixes #231