Add a react 18 preload namespace
React 18 introduces a new render API:

https://react.dev/blog/2022/03/29/react-v18#react-dom-client

The reagent.dom.client namespace wraps this new API. reagent.dom still
wraps the old one. React throws a deprecation warning if you use the
old API with React 18.

If a user uses the new API in their project, they will still get the
warning because of re-frame-10x using the old API. This is confusing.
It looks as if their code is suspect, when actually ours is.

re-frame-10x depends on the user's react version. That means we can't
predict whether the user will use React 17 or React 18, or which API
they will invoke. We can't just load reagent.dom.client, because it
only works with React 18. We can't load it conditionally based on the
react version, since it's a general limitation of cljs that we can't
require a namespace at runtime.

Added a second preload namespace. Since only the user can know which
API to use, we let them declare it themselves. The original preload
remains unchanged:

{:preloads [day8.re-frame-10x.preload]}

It calls the old API, which works in all cases. It throws the
deprecation warning if the user has React 18.

The new preload only works with React 18, using the new API properly:

{:preloads [day8.re-frame-10x.preload.react-18]}

Fixes #367