Make path annotations nil-safe
Adding and deleting app-db path inspector pods caused a crash related
to null target DOM refs. I think React started to handle these refs
differently when we upgraded to version 18.

Added nil checks so that it doesn't try to call methods & fields on
nil.

This component could use a more involved refactor, but for now it
seems to work again.