Tag sorted maps when serializing app-db for editing.
Sorted maps are a common datatype which clojure.edn doesn't handle. It
just serializes sorted maps as regular maps.

Reading this back in later causes a crash in our todomvc, since
todomvc.db enforces the type:

https://github.com/day8/re-frame-10x/blob/e294ccf6c430d105f8117ca5e42bf3ea318bb171/examples/todomvc/src/todomvc/db.cljs#L27
https://github.com/day8/re-frame-10x/blob/e294ccf6c430d105f8117ca5e42bf3ea318bb171/examples/todomvc/src/todomvc/db.cljs#L81

Serializing to localstore works, but only because we explicitly
convert that subtree whenever we import the app-db. So there's no
general solution.

https://github.com/day8/re-frame-10x/blob/e294ccf6c430d105f8117ca5e42bf3ea318bb171/examples/todomvc/src/todomvc/db.cljs#L81

Now that re-frame-10x is serializing the app-db for its own
purposes (editing and import/export), it would be nice if it could
just handle sorted maps without secretly turning them into regular
maps.

Now we read and write a #re-frame-10x/sorted-map tag for this editing
feature. Sorted maps are preserved.